requirements.txt
delta==0.4.2
flask==1.1.1
spacy==2.2.3
kafka-python==2.0.1
textblob==0.15.3
plotly==4.1.1
psycopg2==2.8.3
pandas==0.25.1
numpy
Flask-Menu==0.7.1
flask-paginate==0.5.5
tweepy==3.8.0
pyspark==2.4.4
pyarrow==0.16.0
#EDA
python-snappy==0.5.4
wordcloud==1.4.1
pillow==7.0.0
imageio==2.6.1
textacy==0.10.0
tqdm
swifter==0.301
pytest==5.3.5
fastparquet==0.3.3
gin-config==0.3.0
sqlalchemy==1.3.8
# DL
tensorflow
dash==1.9.1
pytest-spark<0.6.0
pytest-flask==1.0.0
coverage==5.0.4
snorkel==0.9.3
absl-py==0.9.0
#documentation
sphinx==1.0.3
sphinx_rtd_theme==3.0.1
sphinxcontrib-bibtex==1.0.1
recommonmark==0.6.0
m2r==0.2.1
sphinx-copybutton==0.2.10
sphinx-argparse==0.2.5

